requirements.txt
attrs==19.3.0
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.4.1
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.10.1
decorator==4.4.2
docutils==0.15.2
fonttools==4.31.2
Fiona==1.8.20
geographiclib==1.52
geopy==2.2.0
geopandas==0.8.1
greenlet==1.1.2
idna==2.9
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.3.2
matplotlib==3.5.1
munch==2.5.0
networkx==2.4
numpy==1.21.4
olefile==0.46
packaging==21.3
pandas==1.3.4
psycopg2-binary==2.9.2
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==3.3.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.3
PyWavelets==1.1.1
s3transfer==0.3.3
Shapely==1.8.0
six==1.14.0
snuggs==1.4.7
SQLAlchemy==1.4.27
toolz==0.10.0
tornado==6.0.4
urllib3==1.25.8